feat(e2e): Api 34 for android e2e tests
