+++ suite-native/app/e2e/pageObjects/accountImportActions.ts
@@ -28,7 +28,25 @@ class AccountImportActions {
         const coinItemElement = element(
             by.id(`@onboarding/select-coin/${networkSymbol.toLowerCase()}`),
         );
-        await scrollUntilVisible(coinItemElement);
+        // Espresso's native search-scroll loop (waitFor+whileElement+scroll) fires rapid scroll
+        // events that flood the JS bridge on API 34 with the non-virtualized coin list, causing
+        // the JS thread to become unresponsive (ANR) and the app to crash. A JS-controlled loop
+        // with async/await boundaries lets the event loop drain scroll callbacks between steps.
+        //
+        // 1500 px per step ensures coins deep in the list (e.g. ZEC at y≈3066 px) become visible
+        // in a single scroll. The first scroll triggers the ScrollDivider FadeIn (Reanimated), which
+        // temporarily elevates swiftshader memory usage. A second scroll on top of that elevated
+        // baseline OOM-kills the QEMU emulator process on memory-constrained CI runners. One large
+        // scroll avoids a second scroll entirely.
+        const scrollView = element(by.id('@screen/mainScrollView'));
+        for (let i = 0; i < 5; i++) {
+            try {
+                await detoxExpect(coinItemElement).toBeVisible(75);
+                break;
+            } catch {
+                await scrollView.scroll(1500, 'down');
+            }
+        }
         await coinItemElement.tap();
         await detoxExpect(element(by.id('@screen/XpubScan'))).toBeVisible();
     }
